+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.14.0
 
 - @iov/bcp: Add `AtomicSwapHelpers.createId`.
+- @iov/bns: Cache token data from chain
 
 Breaking changes
 
@@ -21,6 +22,9 @@ Breaking changes
   `TransactionQuery`.
 - @iov/bcp: Rename `BcpAtomicSwapConnection` to `AtomicSwapConnection` and
   `BcpConnection` to `BlockchainConnection`.
+- @iov/bcp: Rename `BcpTicker` to `Token`.
+- @iov/bcp: Rename `BcpConnection.getTicker` to `getToken` and
+  `BcpConnection.getAllTickers` to `.getAllTokens`.
 
 ## 0.13.3